config/python/theme/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "theme"
version = "1.0.0"
description = ""
requires-python = ">=3.12"
dependencies = [
    "pyyaml",
]